+++ app/src/main/java/com/pasich/mynotes/extendedEditor/models/ParsedNote.java
@@ -33,7 +33,7 @@ public class ParsedNote {
             return new Gson().toJson(attachments);
         } catch (Exception e) {
             Log.e("jsonToModel", "Failed to convert attachments to JSON", e);
-            return "[]"; // безпечний fallback
+            return "[]";
         }
     }