+++ app/src/main/java/com/pasich/mynotes/ui/view/dialogs/main/AllTagSelectDialog.java
@@ -2,7 +2,9 @@ package com.pasich.mynotes.ui.view.dialogs.main;
 
 import android.content.Context;
 import android.view.LayoutInflater;
+import android.view.View;
 import android.widget.ScrollView;
+import android.widget.TextView;
 
 import androidx.appcompat.app.AlertDialog;